feat : RecipeSubIngredient 관련 Controller 수정(api, ResponseEntity)

+++ molaeng_back/src/main/java/idle/molaeng_back/recipe/model/response/RecipeIngredientRes.java
@@ -8,11 +8,11 @@ import lombok.Getter;
 public class RecipeIngredientRes {
     long ingredientId;
     String ingredientName;
-    //필요무게가 erd랑 entity에는 int인데 실제 데이터는 소수점도 있는데 -> float로 수정
-    float needWeight;
+    //필요무게가 erd랑 entity에는 int인데 실제 데이터는 소수점도 있는데 -> double로 수정
+    double needWeight;
     String weightUnit;
 
-    public RecipeIngredientRes(long ingredientId, String ingredientName, float needWeight, String weightUnit) {
+    public RecipeIngredientRes(long ingredientId, String ingredientName, double needWeight, String weightUnit) {
         this.ingredientId = ingredientId;
         this.ingredientName = ingredientName;
         this.needWeight = needWeight;
